refactor index pkg around the concept of indexTree
The HTML index is generated in two conceptual steps:

1. make an idexTree out of a directory structure.
   This indexTree is an intermediate representation of
   the index; and
2. generate an HTML index from the indexTree representation.

This refactor aims at making this idea more explicit throughout
the package and fixing inconsistencies

- rename tests, functions, arguments, varialbes
- split indexTree related code in its own file and test file
- adding docstrings
- misc style fixes
