+++ index/utils_test.go
@@ -1,7 +1,10 @@
 package index
 
-import "testing"
-import "github.com/stretchr/testify/assert"
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
 
 func Test_toTitle(t *testing.T) {
 	actual := toTitle("my_page.md")
